feature(brew): generate `Brewfile` (#16)
Add `Brewfile` to track Homebrew formulae and casks to install across
machines. Host-specific software are determined by checking the
machine's hostname with `hostname -s`. Note that the `-s` command is
only available on BSD systems.

To install the software specified in `Brewfile`, cd into the home
directory (or wherever `stow` symlinks the dotfiles) and run `brew
bundle install`. Once successful, it will generate `Brewfile.lock.json`
in the same directory -- there is no need to commit this file as it is
only used by `brew bundle` for troubleshooting.

Since Homebrew is a rolling release package manager, it does not support
installing arbitrary older versions of software. As a result, `Brewfile`
only informs Homebrew to install the latest versions of the listed
formulae and casks.