chore(git): define Git `signingkey` for UTS repositories (#24)
Define `signingkey` and other user properties for UTS GitLab
repositories. This configuration will be applied to all machines.
Perhaps it would be nice to figure out how to load this file on specific
machines, or even change the `signingkey`, as detailed in issue #23.

Add header comment in both Git configuration files so Vim and GitHub
applies appropriate syntax highlighting.