chore(kitty): fix Nerd Font symbol map
Kitty was displaying the wrong icon for the branch icon on Ubuntu (it
was displaying an angle icon). I'm unsure if this was the case for other
icons. Regardless, I've created this patch to fix this issue.

Taken from Kitty's FAQ page, this patch manually defines the symbol map
for all the codepoints used by Nerd Font Mono. As evident from the
comment, this patch defines symbol mappings from `v3.1.0`, however, the
installed version on my machine is `v3.2.0`. I'm unsure if there are
breaking changes regarding the mappings between these two versions, but
it seems to have fixed the issue for now.