fix(zsh): override `SHORT_HOST` before host-specific env vars
Defining `SHORT_HOST` in `.profile` means that it would be overridden by
Oh My Zsh in `.zshrc`. To avoid this, I defined `SHORT_HOST` in `.zshrc`
after the `source` command and moved all the host-specific variables to
the "User configuration" section. I also simplified the `SHORT_HOST`
value as I will not be using the `ComputerName` property from macOS's
`scutil`.